app_detail_screenをリファクタリング
